pyproject.toml
[project]
name = "events_ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "devtools>=0.12.2",
    "fpdf2>=2.8.4",
    "google-genai>=1.30.0",
    "htmlcorder",
    "icalendar>=6.3.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "lxml>=5.3.1",
    "markdownify>=1.2.0",
    "moviepy@git+https://github.com/Zulko/moviepy",
    "nicegui>=2.24.2",
    "ollama>=0.4.7",
    "openai>=2.6.1",
    "pandas>=2.2.3",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.3",
    "selenium>=4.35.0",
    "structlog>=25.4.0",
]

[project.scripts]
main = "events_ai:main_cli"
heygen = "events_ai.agents.heygen_client:heygen_cli"

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "devtools>=0.12.2",
    "pytest>=9.0.2",
]

[tool.uv.sources]
htmlcorder = { git = "https://github.com/superlou/htmlcorder.git" }
